pyproject.toml
[project]
name = "name"
version = "0.1.0"
description = ""
authors = [
    {name = "danila",email = "danniila2000@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "envparse (>=0.2.0,<0.3.0)",
    "supervisor (>=4.2.5,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "ruff (>=0.11.11,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
